+++ mlflow/dspy/autolog.py
@@ -1,4 +1,5 @@
 from mlflow.dspy.save import FLAVOR_NAME
+from mlflow.dspy.util import save_dspy_module_state
 from mlflow.tracing.provider import trace_disabled
 from mlflow.utils.annotations import experimental
 from mlflow.utils.autologging_utils import (
@@ -13,6 +14,7 @@ def autolog(
     log_traces: bool = True,
     log_traces_from_compile: bool = False,
     log_traces_from_eval: bool = True,
+    log_compiles: bool = False,
     disable: bool = False,
     silent: bool = False,
 ):
@@ -30,6 +32,8 @@ def autolog(
             `built-in evaluator <https://dspy.ai/learn/evaluation/metrics/#evaluation>`_.
             If ``False``, traces are only logged from normal model inference and disabled when
             running the evaluator. Default to ``True``.
+        log_compiles: If ``True``, information about the optimization process is logged when
+            `Teleprompter.compile()` is called.
         disable: If ``True``, disables the DSPy autologging integration. If ``False``,
             enables the DSPy autologging integration.
         silent: If ``True``, suppress all event logs and warnings from MLflow during DSPy
@@ -45,6 +49,7 @@ def autolog(
         log_traces=log_traces,
         log_traces_from_compile=log_traces_from_compile,
         log_traces_from_eval=log_traces_from_eval,
+        log_compiles=log_compiles,
         disable=disable,
         silent=silent,
     )
@@ -54,7 +59,7 @@ def autolog(
     from mlflow.dspy.callback import MlflowCallback
 
     # Enable tracing by setting the MlflowCallback
-    if log_traces and not disable:
+    if not disable:
         if not any(isinstance(c, MlflowCallback) for c in dspy.settings.callbacks):
             dspy.settings.configure(callbacks=[*dspy.settings.callbacks, MlflowCallback()])
     else:
@@ -63,9 +68,38 @@ def autolog(
         )
 
     # Patch teleprompter/evaluator not to generate traces by default
-    def trace_disabled_fn(original, self, *args, **kwargs):
+    def patch_fn(original, self, *args, **kwargs):
         # NB: Since calling mlflow.dspy.autolog() again does not unpatch a function, we need to
         # check this flag at runtime to determine if we should generate traces.
+        @trace_disabled
+        def _trace_disabled_fn(self, *args, **kwargs):
+            return original(self, *args, **kwargs)
+
+        def _compile_fn(self, *args, **kwargs):
+            if callback := _active_callback():
+                callback.within_compile = True
+            try:
+                if get_autologging_config(FLAVOR_NAME, "log_traces_from_compile"):
+                    result = original(self, *args, **kwargs)
+                else:
+                    result = _trace_disabled_fn(self, *args, **kwargs)
+                return result
+            finally:
+                if callback:
+                    callback.within_compile = False
+
+        if isinstance(self, Teleprompter):
+            if not get_autologging_config(FLAVOR_NAME, "log_compiles"):
+                return _compile_fn(self, *args, **kwargs)
+
+            # TODO: we may want to save the trainset with mlflow.log_input()
+            program = _compile_fn(self, *args, **kwargs)
+            # Save the state of the best model in json format
+            # so that users can see the demonstrations and instructions.
+            save_dspy_module_state(program, "best_model.json")
+
+            return program
+
         if isinstance(self, Teleprompter) and get_autologging_config(
             FLAVOR_NAME, "log_traces_from_compile"
         ):
@@ -76,11 +110,7 @@ def autolog(
         ):
             return original(self, *args, **kwargs)
 
-        @trace_disabled
-        def _fn(self, *args, **kwargs):
-            return original(self, *args, **kwargs)
-
-        return _fn(self, *args, **kwargs)
+        return _trace_disabled_fn(self, *args, **kwargs)
 
     from dspy.evaluate import Evaluate
     from dspy.teleprompt import Teleprompter
@@ -95,7 +125,8 @@ def autolog(
                 FLAVOR_NAME,
                 cls,
                 compile_patch,
-                trace_disabled_fn,
+                patch_fn,
+                manage_run=get_autologging_config(FLAVOR_NAME, "log_compiles"),
             )
 
     call_patch = "__call__"
@@ -104,7 +135,7 @@ def autolog(
             FLAVOR_NAME,
             Evaluate,
             call_patch,
-            trace_disabled_fn,
+            patch_fn,
         )
 
 
@@ -117,9 +148,20 @@ def _autolog(
     log_traces: bool,
     log_traces_from_compile: bool,
     log_traces_from_eval: bool,
+    log_compiles: bool,
     disable: bool = False,
     silent: bool = False,
 ):
     """
-    TODO: Implement patching logic for autologging models and artifacts.
+    TODO: Implement patching logic for autologging artifacts.
     """
+
+
+def _active_callback():
+    import dspy
+
+    from mlflow.dspy.callback import MlflowCallback
+
+    for callback in dspy.settings.callbacks:
+        if isinstance(callback, MlflowCallback):
+            return callback