+++ mlflow/server/js/src/experiment-tracking/components/run-page/RunViewOverview.tsx
@@ -19,6 +19,7 @@ import { RunViewTagsBox } from './overview/RunViewTagsBox';
 import { RunViewDescriptionBox } from './overview/RunViewDescriptionBox';
 import { DetailsOverviewMetadataRow } from '../DetailsOverviewMetadataRow';
 import { RunViewRegisteredModelsBox } from './overview/RunViewRegisteredModelsBox';
+import { RunViewRegisteredPromptsBox } from './overview/RunViewRegisteredPromptsBox';
 import { RunViewLoggedModelsBox } from './overview/RunViewLoggedModelsBox';
 import { RunViewSourceBox } from './overview/RunViewSourceBox';
 import { DetailsOverviewMetadataTable } from '@mlflow/mlflow/src/experiment-tracking/components/DetailsOverviewMetadataTable';
@@ -60,6 +61,18 @@ export const RunViewOverview = ({
 
   const registeredModelVersionSummaries = registeredModelVersionSummariesForRun;
   const shouldDisplayLoggedModelsBox = loggedModels?.length > 0;
+
+  const renderPromptMetadataRow = () => (
+    <DetailsOverviewMetadataRow
+      title={
+        <FormattedMessage
+          defaultMessage="Registered prompts"
+          description="Run page > Overview > Run prompts section label"
+        />
+      }
+      value={<RunViewRegisteredPromptsBox runUuid={runUuid} />}
+    />
+  );
 
   const renderDetails = () => {
     return (
@@ -173,6 +186,7 @@ export const RunViewOverview = ({
             )
           }
         />
+        {renderPromptMetadataRow()}
       </DetailsOverviewMetadataTable>
     );
   };